ferry_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
python-multipart==0.0.6

# Existing dependencies
python-dotenv==1.0.0
chromadb==0.4.18
google-generativeai==0.3.2
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0

# Multi-Agent Framework
langgraph==0.0.69
langchain-core==0.1.52
langchain-google-genai==1.0.10
langchain-community==0.0.38

# Communication and Caching
redis==5.0.1
aioredis==2.0.1

# Monitoring and Metrics
prometheus-client==0.19.0

# Additional utilities
asyncio==3.4.3
